- refactored solution
